proper definition and word being used in Hangman
